Added location base
