+++ .idea/misc.xml
@@ -9,6 +9,7 @@
         <entry key="../../../../../layout/compose-model-1644918686739.xml" value="0.32559121621621623" />
         <entry key="../../../../../layout/compose-model-1644919494445.xml" value="0.28055555555555556" />
         <entry key="../../../../../layout/compose-model-1644919781084.xml" value="0.1" />
+        <entry key="../../../../../layout/compose-model-1645084132505.xml" value="0.41834451901565994" />
         <entry key="../../../../layout/compose-model-1644928077276.xml" value="0.14611204013377926" />
         <entry key="../../../../layout/compose-model-1644928632733.xml" value="0.2951858108108108" />
         <entry key="../../../../layout/compose-model-1645015855717.xml" value="0.4865771812080537" />
