+++ app/build.gradle
@@ -68,4 +68,7 @@ dependencies {
 
     // Graph
     implementation 'com.github.PhilJay:MPAndroidChart:v3.1.0'
+
+    // Google location
+    implementation 'com.google.android.gms:play-services-location:19.0.1'
 }
