+++ app/src/main/java/fi/sportionbois/sportion/composables/Settings.kt
@@ -1,0 +1,9 @@
+package fi.sportionbois.sportion.composables
+
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+
+@Composable
+fun Settings(name: String) {
+    Text(text = name)
+}
