+++ app/src/main/java/fi/sportionbois/sportion/location/LocationHandler.kt
@@ -1,0 +1,41 @@
+package fi.sportionbois.sportion.location
+
+import android.Manifest
+import android.content.Context
+import android.content.pm.PackageManager
+import android.util.Log
+import androidx.core.app.ActivityCompat
+import com.google.android.gms.location.*
+
+class LocationHandler() {
+    private lateinit var fusedLocationClient: FusedLocationProviderClient
+    private lateinit var locationCallback: LocationCallback
+
+    fun initializeLocation(context: Context) {
+        fusedLocationClient = LocationServices.getFusedLocationProviderClient(context)
+        if (ActivityCompat.checkSelfPermission(
+                context,
+                Manifest.permission.ACCESS_FINE_LOCATION
+            ) ==
+            PackageManager.PERMISSION_GRANTED
+        ) {
+            fusedLocationClient.lastLocation.addOnSuccessListener {
+                Log.d(
+                    "GEOLOCATION",
+                    "last location latitude: ${it?.latitude} and longitude: ${it?.longitude}"
+                )
+            }
+        }
+        locationCallback = object : LocationCallback() {
+            override fun onLocationResult(p0: LocationResult) {
+                p0 ?: return
+                for (location in p0.locations) {
+                    Log.d(
+                        "GEOLOCATION",
+                        "new location latitude: ${location.latitude} and longitude: ${location.longitude}"
+                    )
+                }
+            }
+        }
+    }
+}
